+++ include/all.h
@@ -53,6 +53,7 @@
 #include "click.h"
 #include "key_press.h"
 #include "floating.h"
+#include "drag.h"
 #include "configuration.h"
 #include "handlers.h"
 #include "randr.h"
